fix: Add Suspense boundary for useSearchParams hook
