+++ apps/client/src/app/order/form/OrderFormPageContent.tsx
@@ -9,13 +9,13 @@ import PriceSummarySection from '@/components/features/order/form/PriceSummarySe
 import AgreementSection from '@/components/features/order/form/TermsSection';
 import OrderItemsSection from '@/components/features/order/OrderItemsSection';
 import Header from '@/components/layouts/Header';
-import { OrderFrom } from '@/constants/order';
 import { ORDER_TERMS } from '@/constants/terms';
 import { useTermsCheck } from '@/hooks/check';
 import { useAddressListQuery } from '@/hooks/queries/useAddressQuery';
 import { useBaseOrderRequestStore } from '@/hooks/stores/useBaseOrderRequestStore';
 import { useSelectedAddressIdStore } from '@/hooks/stores/useSelectedAddressIdStore';
 import { useSelectedProductsStore } from '@/hooks/stores/useSelectedProductsStore';
+import { useValidOrderFrom } from '@/hooks/useValidOrderFrom';
 import { AddressInfo } from '@/schemas/address';
 import { formatPrice } from '@/utils/formatPrice';
 import { getPriceSummary } from '@/utils/order/getPrice';
@@ -26,38 +26,26 @@ export type DeliveryRequest = {
     customText: string;
 };
 
-export default function OrderFormPageContent({
-    orderFrom,
-}: {
-    orderFrom: OrderFrom;
-}) {
+export default function OrderFormPageContent() {
     const router = useRouter();
-
+    const { orderFrom, isValid } = useValidOrderFrom();
+    // 모든 훅은 여기서 호출
     const [activeAddressInfo, setActiveAddressInfo] = useState<AddressInfo>();
     const { selectedAddressId, setSelectedAddressId } =
         useSelectedAddressIdStore();
-
-    // deliveryRequest로 네이밍 변경
     const [deliveryRequest, setDeliveryRequest] = useState<DeliveryRequest>({
         option: null,
         customText: '',
     });
-
-    // 약관 체크 상태 관리
     const termsCheck = useTermsCheck(ORDER_TERMS);
     const isAllTermsChecked = termsCheck.isAllMandatoryChecked;
+    const { products } = useSelectedProductsStore();
+    const { productPrice, totalPrice } = getPriceSummary(products);
+    const { setBaseOrderRequest } = useBaseOrderRequestStore();
+    const { data: addressList, isLoading, error } = useAddressListQuery();
 
     const isDisabled = !(activeAddressInfo && isAllTermsChecked);
 
-    const { products } = useSelectedProductsStore();
-    const { productPrice, totalPrice } = getPriceSummary(products);
-
-    const { setBaseOrderRequest } = useBaseOrderRequestStore();
-
-    // 배송지 목록 쿼리
-    const { data: addressList, isLoading, error } = useAddressListQuery();
-
-    // 페이지 진입 시 기본 배송지 설정
     useEffect(() => {
         const info =
             !addressList || addressList.length === 0
@@ -79,39 +67,35 @@ export default function OrderFormPageContent({
         }
     }, [addressList, selectedAddressId, setSelectedAddressId]);
 
-    // Delivery request handler
     const onDeliveryRequestChange = <K extends keyof DeliveryRequest>(
         key: K,
         value: DeliveryRequest[K]
     ) => {
-        setDeliveryRequest((prev) => ({
-            ...prev,
-            [key]: value,
-        }));
+        setDeliveryRequest((prev) => ({ ...prev, [key]: value }));
     };
 
-    // 결제 버튼 클릭 핸들러 분리
+    if (!isValid) {
+        return <div>잘못된 접근입니다. 홈으로 이동합니다...</div>;
+    }
+
     const handleOrderButtonClick = () => {
         if (!activeAddressInfo) return;
-
         const request =
             deliveryRequest.option === '직접 입력'
                 ? deliveryRequest.customText
                 : deliveryRequest.option;
-
         setBaseOrderRequest({
             addressId: activeAddressInfo.addressId,
             deliveryRequest: request ?? '',
             agreedToTerms: isAllTermsChecked,
         });
-
+        // isValid가 true → orderFrom은 OrderFrom 타입임이 보장됨
         router.push(`/order/confirm?${getOrderFromQueryString(orderFrom)}`);
     };
 
     return (
         <div className="pb-17">
             <Header type="title" title="주문/결제" />
-
             <div className="flex flex-col px-5 pt-2">
                 {/* 배송지 정보 */}
                 <section className="flex flex-col gap-3 border-t border-black pb-8 pt-4">
@@ -125,12 +109,10 @@ export default function OrderFormPageContent({
                         error={error}
                     />
                 </section>
-
                 {/* 주문 상품 */}
                 <section className="flex flex-col gap-3 border-t border-black pb-8 pt-4">
                     <OrderItemsSection variant="form" />
                 </section>
-
                 {/* 결제 금액 */}
                 <section className="flex flex-col border-t border-black pb-5 pt-5">
                     <PriceSummarySection
@@ -138,13 +120,11 @@ export default function OrderFormPageContent({
                         totalPrice={totalPrice}
                     />
                 </section>
-
                 {/* 약관 동의 */}
                 <section className="flex flex-col border-t border-black pb-5 pt-5">
                     <AgreementSection termsCheck={termsCheck} />
                 </section>
             </div>
-
             {/* 버튼 */}
             <BottomButton
                 text={`${formatPrice(totalPrice)} 원 결제하기`}
