+++ apps/client/src/components/common/SuspenseWrapper.tsx
@@ -1,0 +1,12 @@
+'use client';
+import { ReactNode, Suspense } from 'react';
+
+export default function SuspenseWrapper({
+    children,
+    fallback = 'Loading...',
+}: {
+    children: ReactNode;
+    fallback?: ReactNode;
+}) {
+    return <Suspense fallback={fallback}>{children}</Suspense>;
+}
